+++ src/MainComponent.js
@@ -1,0 +1,18 @@
+// import React from "react";    
+// import TodoList from "./TodoList";
+// import AddTodoForm from "./AddTodoForm";
+                     
+// const MainComponent = ({addTodo, isLoading, removeTodo, todoList, editTitle, updateAirtableRecord}) => {
+//         return(
+//             <>
+//               <h1>Todo List</h1>
+//               <AddTodoForm onAddTodo={addTodo} />
+//               {/* <Search onSearch={onSearch} /> */}
+//               {isLoading ? (
+//                 <p>Loading...</p>
+//               ) : (
+//                 <TodoList onRemoveTodo={removeTodo} todoList={todoList} editTitle={editTitle} onSaveTodo={updateAirtableRecord} />
+//              )}
+//             </>)
+// }
+// export default MainComponent;
